# Conflicts:
#	src/main/java/edu/xpu/cs/lovexian/app/appadmin/Kafka/KafkaReceiver.java
#	src/main/java/edu/xpu/cs/lovexian/app/appadmin/service/IDeviceStatisticsAdminService.java
#	src/main/java/edu/xpu/cs/lovexian/app/appadmin/service/ISensorsDataAdminService.java
#	src/main/java/edu/xpu/cs/lovexian/app/appadmin/service/impl/DeviceStatisticsAdminServiceImpl.java
#	src/main/java/edu/xpu/cs/lovexian/app/appadmin/service/impl/SensorsDataAdminServiceImpl.java
#	target/classes/application.yml

+++ src/main/java/edu/xpu/cs/lovexian/common/utils/TransferVUtil.java
@@ -24,7 +24,7 @@ public class TransferVUtil {
 
 	private static int currentP;
 
-	public static int encrypt(double currentV) {
+	public static int encrypt(float currentV) {
 		if(currentV<=12600&&currentV>=9300) {
 			for (int i = 0; i < 11; i++) {
 				if (currentV >= Vms[i] && currentV < Vms[i - 1]) {
